+++ packages/react/src/components/date-picker/date-picker.react-template.tsx
@@ -9,7 +9,18 @@ export const reactDatePicker: ComponentImplementation<DatePicker> = {
   component: "datePicker",
   implementation: "react",
   template: () =>
-    function datePickerTemplate({ id, dsoDateChange, value, min, max, disabled, invalid, autofocus, direction }) {
+    function datePickerTemplate({
+      id,
+      dsoDateChange,
+      value,
+      min,
+      max,
+      disabled,
+      invalid,
+      autofocus,
+      direction,
+      describedBy,
+    }) {
       return (
         <DsoDatePicker
           onDsoDateChange={(e: CustomEvent<DsoDatePickerChangeEvent>) => dsoDateChange?.(e)}
@@ -21,6 +32,7 @@ export const reactDatePicker: ComponentImplementation<DatePicker> = {
           dsoAutofocus={autofocus}
           disabled={disabled}
           invalid={invalid}
+          describedBy={describedBy}
         ></DsoDatePicker>
       );
     },